+++ development/settings.py
@@ -33,7 +33,7 @@ POSTGRES_PORT = int(os.environ.get('POSTGRES_PORT', 5432))
 
 DATABASES = {
     'default': {
-        'ENGINE': 'django.db.backends.postgresql_psycopg2',
+        'ENGINE': 'django.contrib.gis.db.backends.postgis',
         'NAME': POSTGRES_DB,
         'USER': POSTGRES_USER,
         'PASSWORD': POSTGRES_PASSWORD,